test(core): split vite plugin loader tests
